test: move exports test, remove cjs test
